Moved to proper handeling of unicode.
Due to a lack of python knowledge this plugin was dumbly adding ascii and non ascii characters. This was leading to a potential of horribly garbled results when more than one non ascii character was inserted in a row.

Now the plugin takes in text as the hexchat reported charset, handles everything internally as unicode strings and returns it back in the reported charset. 

This should hopefully fix any issues with horribly garbled results.