Added version that works in Python 3.x
The original file didn't work in Python 3.x (caused errors), so I've
made a version that works in 3.x. It has less code than the original
version, so I'm guessing this version wouldn't work in Python 2.x. So, I
think it's best to make it 2 separate files (like I did).